Currency Converter
The provided Java program implements a currency converter that allows users to convert between US Dollars, LKR Rupees, INR Rupees, and Pounds. It prompts the user to select the current currency and the currency they want to convert to, as well as input the amount to convert. The program then performs the conversion based on the latest currency values and displays the converted amount. The currency values are last updated on 8/2/2024. But you could use an API to get current currency rates rather than changing it everyday.